requirements.txt
PyQt6>=6.6,<7
pywinauto>=0.6.8,<1
sentence-transformers>=4,<5
PyYAML>=6,<7
spacy>=3.7,<4
chromadb>=0.6,<1
yake>=0.7,<1
pdfplumber>=0.10,<1
python-docx>=1.0,<2
python-pptx>=0.6,<1
